Programowanie/Praca_magisterska/requirements.txt
opencv-python
Pillow
matplotlib
pyserial
numpy